+++ packages/analytics-js/__mocks__/remotePlugins/IubendaConsentManager.ts
@@ -1,0 +1,10 @@
+const IubendaConsentManager = () => ({
+  name: 'IubendaConsentManager',
+  consentManager: {
+    init: jest.fn(() => {}),
+    updateConsentsInfo: jest.fn(() => {}),
+    isDestinationConsented: jest.fn(() => {}),
+  },
+});
+
+export default IubendaConsentManager;