+++ packages/analytics-js-cookies/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.3.12](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.3.11...@rudderstack/analytics-js-cookies@0.3.12) (2024-10-21)
+
+### Dependency Updates
+
+* `@rudderstack/analytics-js-common` updated to version `3.10.0`
 ## [0.3.11](https://github.com/rudderlabs/rudder-sdk-js/compare/@rudderstack/analytics-js-cookies@0.3.10...@rudderstack/analytics-js-cookies@0.3.11) (2024-10-18)
 
 ### Dependency Updates